requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dropbox==11.36.2
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.103.1
Flask==2.3.3
Flask-Login==0.6.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
passlib==1.7.4
pdfkit==1.0.0
Pillow==10.0.0
ply==3.11
pyasn1==0.5.0
pydantic==2.3.0
pydantic_core==2.6.3
PyMySQL==1.1.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
stone==3.3.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
Werkzeug==2.3.7
